Migrating subscriber info to plugin architecture
