+++ server/routes.js
@@ -1,6 +1,6 @@
 var OpenTok = require('opentok');
 
-module.exports = function (app, config, redis, ot, anvil, redirectSSL) {
+module.exports = function (app, config, redis, ot, redirectSSL) {
   var RoomStore = require('./roomstore.js')(redis, ot);
   app.get('*', function(req, res, next) {
     if (req.host === 'hangout.tokbox.com') {
@@ -83,39 +83,6 @@ module.exports = function (app, config, redis, ot, anvil, redirectSSL) {
           chromeExtensionId: config.chromeExtensionId
         });
       }
-    });
-  });
-
-  app.get('/:room/subscriber/:subscriberId', function(req, res) {
-    var room = req.param('room');
-    var subscriberId = req.param('subscriberId');
-
-    RoomStore.getRoom(room, function(err, sessionId, apiKey, secret) {
-      if (err) {
-        console.error('Error getting room: ', err);
-        res.send({ error: err.message });
-        return;
-      } else if(!apiKey || !secret) {
-        apiKey = ot.apiKey;
-        secret = ot.apiSecret;
-      }
-
-      var payload = {
-        apiKey: apiKey,
-        apiSecret: secret,
-        sessionId: sessionId,
-        subscriberId: subscriberId
-      };
-
-      anvil.getSubscriberInfo(payload, function(err, info) {
-        if (err) {
-          console.error('Error retrieving subscriber information: ', err);
-          res.send({ error: err.message });
-          return;
-        }
-
-        res.send({info: info});
-      });
     });
   });
